+++ config/config.go
@@ -23,7 +23,7 @@ const (
 	AllowedOrigins   = "allowed.origins"
 	SendGridAPIKey   = "sendgrid.api.key"
 	DomainName       = "domain.name"
-	PapertrailAddr   = "papertrail.addr"
+	SyslogAddr       = "syslog.addr"
 )
 
 // Config defines a configuration abstraction so we don't rely on any specific